Check that zsh, tmux and emacs are installed

+++ bootstrap.sh
@@ -1,6 +1,25 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 
 set -e
+
+command_exists() {
+	command -v "$@" >/dev/null 2>&1
+}
+
+if ! command_exists zsh; then
+    echo "Error: zsh is not installed!"
+    exit 1
+fi
+
+if ! command_exists tmux; then
+    echo "Error: tmux is not installed!"
+    exit 1
+fi
+
+if ! command_exists emacs; then
+    echo "Error: emacs is not installed!"
+    exit 1
+fi
 
 # Remove old cofigurations
 rm -f ~/.zshrc ~/.tmux.conf ~/.emacs
